tools/pyproject.toml
[project]
name = "book-clock-tools"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["freetype-py"]

[tool.uv]
dev-dependencies = []
